grpc-server/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
grpcio-reflection==1.71.0
protobuf==5.29.4
joblib==1.5.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.2.2
torch==2.6.0